Better support for QML and JSON
This commit improves qmlext support for the JSON
format and its interop with QML.

QML wraps objects in QJSValue, that cannot be
converted into JSON properly. This commit "unpacks"
QJSValues into QVariants, so that they can
be transformed into JSON.

This commit also improves testing, adding an
item to bus to item test.